[IMP] website: fixes grammatical errors & updates doc for v14
Reworked the Website portion of the documentation to update the screenshots (and processes) for v14
and condensed (instead of deleted) the easier, more straight-forward portions.

Removed the "Ensure all URLs use custom domain" section of domain_names.rst because I could not find
any section that it spoke about, nor could I find anything called "System Parameters." If there is a
new process (or place to look) for such a process, please feel free to share that process (or place)
with me, and I'll write it into this doc.

I'll need someone with more techinical expertise to confirm that the "GeoIP" doc is correct. Same
for the "Technical hints for customization" section of the Multi-Website doc.