[FIX] odoo_theme: unclickable :show-content: pages in global toctree
When navigating the themes tutorial (developer/howto/themes), the rdtraining page wasn't clickable
in the global toc, even though it was specified as :show-content: (note that the rdtraining was only added in 14+).

The current heuristic extracting the pagename from a toctree node wasn't considering the current depth
if the page referenced by the toctree node was in the same folder than the current page.

Explanation:

We are computing the toctree for the page developer/howtos/themes.
One of the toctree nodes references the rdtraining page, with a relative link,
i.e. `rdtraining.html`.

We were wrongfully looking for show_content in the metadata of the `rdtraining` page,
instead of `developer/howtos/rdtraining`, since we didn't consider the current document
depth when the reference didn't contain any `/`.

This commit ensures that if the current document has a depth > 1, this depth is
currently considered, even for links referencing documents in the same folder.

Task-Id: 2538297